python/requirements.txt
chainlib-eth~=0.4.11
eth-erc20~=0.5.0
funga-eth~=0.6.0
dexif~=0.0.1
